requirements.txt
python=3.10.6
pyspark=3.4.3
jupyterlab=4.2.4
apache-airflow=2.9.3
apache-airflow-providers-apache-spark=4.9.0
apache-superset=4.0.2
